requirements.txt
einops==0.4.1
matplotlib==3.2.2
numpy==1.21.5
tqdm==4.63.0
ptflops==0.6.9
pillow==9.1.1
opencv_python==4.6.0.66
albumentations==1.2.1